# docs/swagger.yaml
basePath: /
definitions:
  cluster.Event:
    properties:
      detail:
        type: string
      id:
        type: string
      serviceId:
        type: string
      status:
        type: string
      timestamp:
        type: string
      title:
        type: string
    type: object
  cluster.Instance:
    properties:
      id:
        type: string
      latency:
        type: integer
      name:
        type: string
      restarts:
        type: integer
      startedAt:
        type: string
      status:
        type: string
    type: object
  cluster.Service:
    properties:
      dependencies:
        items:
          type: string
        type: array
      description:
        type: string
      errorRate:
        type: number
      id:
        type: string
      instances:
        items:
          $ref: '#/definitions/cluster.Instance'
        type: array
      kind:
        type: string
      latency:
        type: integer
      name:
        type: string
      position:
        items:
          type: number
        type: array
      requestsPerMinute:
        type: integer
      status:
        type: string
      uptime:
        type: number
      version:
        type: string
    type: object
  cluster.Snapshot:
    properties:
      events:
        items:
          $ref: '#/definitions/cluster.Event'
        type: array
      generatedAt:
        type: string
      services:
        items:
          $ref: '#/definitions/cluster.Service'
        type: array
    type: object
...
  dto.OAuthCallbackResponse:
    properties:
      email:
        example: student@example.com
        type: string
      message:
        example: успешный вход через OAuth
        type: string
      userId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  handler.AuthErrorResponse:
    properties:
      error:
        example: invalid_or_expired_token
        type: string
    type: object
...
host: localhost:4200
info:
  contact: {}
  description: 'REST API на Go (Gin + MongoDB + Redis + MinIO + RabbitMQ) с JWT/OAuth2-аутентификацией,
    CRUD-ресурсами, файловым хранилищем, асинхронной обработкой событий и Kubernetes-деплоем.
    ЛР9: health-зонды (/health/live, /health/ready), K8s-манифесты, горизонтальное
    масштабирование, Redis distributed lock.'
  title: Lab 2–9 REST API
  version: "1.4"
paths:
  /api/v1/cluster/topology:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Current cluster snapshot
          schema:
            $ref: '#/definitions/cluster.Snapshot'
        "503":
          description: Service Unavailable
          schema:
            type: object
      summary: Получить topology кластера
      tags:
      - cluster
  /api/v1/cluster/services/{serviceID}:
    get:
      produces:
      - application/json
      parameters:
      - description: Идентификатор сервиса
        in: path
        name: serviceID
        required: true
        type: string
      responses:
        "200":
          description: Cluster service
          schema:
            $ref: '#/definitions/cluster.Service'
        "404":
          description: Service not found
          schema:
            type: object
        "503":
          description: Service Unavailable
          schema:
            type: object
      summary: Получить сервис кластера
      tags:
      - cluster
...
  /api/v1/demo/scenarios/{scenario}:
    post:
      produces:
      - application/json
      parameters:
      - description: Сценарий (latency, crash, scale, recover)
        in: path
        name: scenario
        required: true
        type: string
      responses:
        "200":
          description: Updated cluster snapshot
          schema:
            $ref: '#/definitions/cluster.Snapshot'
        "400":
          description: Unknown scenario
          schema:
            type: object
        "503":
          description: Service Unavailable
          schema:
            type: object
      summary: Запустить demo-сценарий
      tags:
      - cluster
...
  /auth/oauth/{provider}:
    get:
      parameters:
      - description: Провайдер (yandex)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Редирект на провайдера
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Инициация OAuth авторизации
      tags:
      - auth
  /auth/oauth/{provider}/callback:
    get:
      parameters:
      - description: Провайдер (yandex)
        in: path
        name: provider
        required: true
        type: string
      - description: Код авторизации
        in: query
        name: code
        required: true
        type: string
      - description: State токен
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: успешный вход через OAuth (JWT в HttpOnly cookies)
          schema:
            $ref: '#/definitions/dto.OAuthCallbackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Обработка OAuth callback
      tags:
      - auth
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"